requirements.txt
tqdm==4.66.1
datasets==2.14.6
transformers==4.35.2
torch==2.1.0
torchsummary==1.5.1
torchvision==0.16.0
Pillow==10.0.1